+++ libs/level1/input/mouse.h
@@ -13,7 +13,7 @@ class Mouse
 {
 public:
 #if PLATFORM == WINDOWS
-	static bool WinProcessMessages(uint32_t message, uint16_t wParam, uint32_t lParam);
+	static bool WinProcessMessages(uint32_t message, uint64_t wParam, uint64_t lParam);
 #endif
 
 	Mouse() : buttonState{} {}
